requirements.txt
uvicorn
fastapi
sqlalchemy
pydantic
python-dotenv
alembic
psycopg2-binary
# required only for development/unit-testing #
pytest
requests
pytest-html
pytest-cov
pytest-dependency
pytest-mock
mock-alchemy
coverage
